enh: first step towards module
